+++ src/app.rs
@@ -388,6 +388,14 @@ impl App {
 
     pub async fn on_message(&mut self, content: Content) -> anyhow::Result<()> {
         // tracing::debug!("incoming: {:#?}", content);
+
+        #[cfg(feature = "dev")]
+        if self.config.developer.dump_raw_messages {
+            if let Err(e) = crate::dev::dump_raw_message(&content) {
+                warn!(error = %e, "failed to dump raw message");
+            }
+        }
+
         let user_id = self.user_id;
 
         let (channel_idx, message) = match (content.metadata, content.body) {
@@ -1162,8 +1170,10 @@ impl App {
     }
 
     fn notify(&self, summary: &str, text: &str) {
-        if let Err(e) = Notification::new().summary(summary).body(text).show() {
-            error!("failed to send notification: {}", e);
+        if self.config.notifications {
+            if let Err(e) = Notification::new().summary(summary).body(text).show() {
+                error!("failed to send notification: {}", e);
+            }
         }
     }
 
@@ -1239,7 +1249,7 @@ impl App {
         self.display_help
     }
 
-    pub(crate) async fn request_contacts_sync(&mut self) -> anyhow::Result<()> {
+    pub async fn request_contacts_sync(&mut self) -> anyhow::Result<()> {
         let now = Utc::now();
         let metadata = self.storage.metadata();
         let do_sync = metadata