+++ src/main.rs
@@ -1,19 +1,10 @@
 //! Signal Messenger client for terminal
 
-mod app;
-mod config;
-mod cursor;
-mod data;
-mod input;
-mod receipt;
-mod shortcuts;
-mod signal;
-mod storage;
-mod ui;
-mod util;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::Arc;
+use std::time::{Duration, Instant};
 
-use app::App;
-
+use clap::Parser;
 use crossterm::{
     event::{
         DisableMouseCapture, EnableMouseCapture, Event as CEvent, EventStream, KeyCode, KeyEvent,
@@ -22,38 +13,40 @@ use crossterm::{
     execute,
     terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
 };
+use gurk::{config, signal, ui};
 use presage::prelude::Content;
-use structopt::StructOpt;
 use tokio_stream::StreamExt;
 use tracing::{error, info, metadata::LevelFilter};
 use tui::{backend::CrosstermBackend, Terminal};
 
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
-use std::time::{Duration, Instant};
-
-use self::storage::JsonStorage;
+use gurk::app::App;
+use gurk::storage::JsonStorage;
 
 const TARGET_FPS: u64 = 144;
 const RECEIPT_TICK_PERIOD: u64 = 144;
 const FRAME_BUDGET: Duration = Duration::from_millis(1000 / TARGET_FPS);
 const SAVE_BUDGET: Duration = Duration::from_millis(1000);
 const RECEIPT_BUDGET: Duration = Duration::from_millis(RECEIPT_TICK_PERIOD * 1000 / TARGET_FPS);
-const MESSAGE_SCROLL_BACK: bool = false;
 
-#[derive(Debug, StructOpt)]
+#[derive(Debug, Parser)]
+#[command(author, version, about, long_about = None)]
 struct Args {
     /// Enables logging to `gurk.log` in the current working directory
-    #[structopt(short, long = "verbose", parse(from_occurrences))]
+    #[clap(short, long = "verbose", action = clap::ArgAction::Count)]
     verbosity: u8,
     /// Relinks the device (helpful when device was unlinked)
-    #[structopt(long)]
+    #[clap(long)]
     relink: bool,
+    /// Dump raw messages to `messages.json` in the current working directory
+    ///
+    /// Used for collecting benchmark data
+    #[clap(long)]
+    dump_messages: bool,
 }
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    let args = Args::from_args();
+    let args = Args::parse();
 
     let file_appender = tracing_appender::rolling::never("./", "gurk.log");
     let (non_blocking, _guard) = tracing_appender::non_blocking(file_appender);