+++ src/storage/mod.rs
@@ -1,11 +1,9 @@
+mod forgetful;
 mod json;
 mod memcache;
 mod storage;
-#[cfg(test)]
-mod test;
 
+pub use forgetful::ForgetfulStorage;
 pub use json::JsonStorage;
 pub use memcache::MemCache;
 pub use storage::{MessageId, Storage};
-#[cfg(test)]
-pub use test::ForgetfulStorage;