+++ CHANGELOG.md
@@ -2,11 +2,23 @@
 
 ## Unreleased (0.3.0-dev)
 
+### Added
+
+- Add notifications config toggling system notifications ([#188], [#192])
+
 ### Changed
 
 - Upgrade presage (this will force relinking the device, due to incompatible changes) ([#182])
 
+### Fixed
+
+-  Use maintenance fixed branch of presage with updated root CA ([#189], [#190])
+
 [#182]: https://github.com/boxdot/gurk-rs/pull/182
+[#188]: https://github.com/boxdot/gurk-rs/pull/188
+[#189]: https://github.com/boxdot/gurk-rs/pull/189
+[#190]: https://github.com/boxdot/gurk-rs/pull/190
+[#192]: https://github.com/boxdot/gurk-rs/pull/192
 
 ## 0.2.5